BPM Detection (experimental) added & many Bugfixes
BPM Detection (start from Preferences Fragment) runs very slow. The
calculation results seems to be not really correctly.
Author of this solution: Peter Backx
(http://www.streamhead.com/easy-bpm-calculation-in-java/). THX to Peter

Bugfixes:
- Check for existing Folder before MusicFolderScanning (avoids crash if
the folder is not correct)

- Track Objects save and return the track duration in milliseconds and
as formatted string now (avoids a lot of crashes in the whole app,
which was depending with the wrong return format of track duration)